Porduct design of desk ATM Sketches version 3
finalisation of the second iteration with description of the inner tech associated to each volumes

a reorientation of the camera Is necessary to facilitate the UX while user will present the screen of their mobile in front of the camera.
indeed the camera needs to dive down aside of the screen, while it display te video feedback for the scanning.

a perpendicular orientation of the camera is useless and unfriendly while using the scan recognition. detection and process.
the user needs to see the screen of its mobile while presenting to the camera.

the screen of the mobile must be horizontal or inclined upside down (screen display up)
while it can face easily the camera lens and the user can guide easily the screen in front of the lens, keeping the mobile device on.